Move new logging code into old file and remove _new file names
